moved login and signup components to pages folder, fixed token references
