+++ src/components/LoggedOutNavbar.js
@@ -9,7 +9,7 @@ const LoggedOutNavbar = () => {
                 <Container>
                     <h1 className={styles.brand}><i className="bi bi-record-circle-fill" style={{ 'color': '#FFD6E0' }}></i> DONUT CAR SHOP</h1>
                     <Nav>
-                        <Nav.Link href="/signin">LOGIN</Nav.Link>
+                        <Nav.Link href="/login">LOGIN</Nav.Link>
                         <Nav.Link href="/signup">SIGN UP</Nav.Link>
                     </Nav>
                 </Container>
